requirements.txt
graphviz==0.16
joblib==1.0.1
numpy==1.20.2
python-dateutil==2.8.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
